etl_service/requirements.txt
pandas
requests
tabulate
python-dotenv
mysql-connector-python
SQLAlchemy
PyYAML